fix: replace long with short in short directories

+++ src/app/(home)/(standard)/(light)/(first-person)/(short)/page.tsx
@@ -3,7 +3,7 @@ import StandardPage from '../../../../../components/pages/standard-page'
 
 export default function Page() {
   const theme: Theme = {
-    verbosity: 'long',
+    verbosity: 'short',
     vibe: 'standard',
     color: 'light',
     tense: 'first-person'
